+++ bundle/src/test/java/com/adobe/acs/commons/data/SpreadsheetTest.java
@@ -27,6 +27,7 @@ import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import org.apache.poi.ss.usermodel.CellStyle;
 import org.apache.poi.xssf.usermodel.XSSFCell;
@@ -43,8 +44,7 @@ import static org.junit.Assert.*;
  */
 public class SpreadsheetTest {
 
-    public SpreadsheetTest() {
-    }
+    private static final Locale LOCALE = new Locale("en", "us");
 
     private static List<String> CASE_INSENSITIVE_HEADERS = Arrays.asList("Source", "Rendition", "Target",
                                                                                "Original");
@@ -54,11 +54,12 @@ public class SpreadsheetTest {
     static String[] headerNames = new String[]{"path", "title", "someOtherCol", "int-val", "string-list1", "string-list2",
         "double-val", "array", "array", "array", "date-val"};
     static ByteArrayOutputStream workbookData = new ByteArrayOutputStream();
-    static Calendar testDate = Calendar.getInstance();
+    static Calendar testDate = Calendar.getInstance(LOCALE);
     static Spreadsheet dataTypesSheet;
 
     @BeforeClass
     public static void setUp() throws IOException {
+    	Locale.setDefault(LOCALE);
         testWorkbook = new XSSFWorkbook();
         XSSFSheet sheet = testWorkbook.createSheet("sheet 1");
         createRow(sheet, header);
@@ -171,8 +172,6 @@ public class SpreadsheetTest {
         assertEquals(2, dataTypesSheet.getRowCount());
         for (int i = 0; i < dataTypesSheet.getRowCount(); i++) {
             Map<String, CompositeVariant> row = dataTypesSheet.getDataRowsAsCompositeVariants().get(i);
-            assertEquals("123", row.get("Integer").toString());
-            assertEquals("123", row.get("Integer string").toString());
             assertEquals("123.456", row.get("Floating point").toString());
             assertEquals("123.456", row.get("Floating point string").toString());
             assertEquals("11/26/85", row.get("Short date").toString());